libgcam_jni: fix manual AE capture aborting
We're bypassing a checker which searches for overexposed
frames from frames payload, and if it detects a frame is
affected by overexposure, it drops the frame, leading to
a forced capture stop if most of the frames are dropped.

Some board have this issue, mostly Exynos is affected by
this. Of course this fix isn't a great one, but by now it
would work fine.